TOP0502 데이터부 입력 처리

+++ t9-trace/src/main/java/rose/mary/trace/database/service/TraceService.java
@@ -16,6 +16,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 
+import pep.per.mint.common.util.Util;
 import rose.mary.trace.core.cache.CacheProxy;
 import rose.mary.trace.core.data.common.State;
 import rose.mary.trace.core.data.common.Trace;
@@ -60,7 +61,7 @@ public class TraceService {
 				// ADD 
 				// 우미건설 요청 (ILIN MI 사이트)
 				// 테스트 미완료
-				if (loadContents) {
+				if (loadContents && !Util.isEmpty(trace.getData())) {					
 					session.delete("rose.mary.trace.database.mapper.m01.TraceMapper.deleteData", trace);
 					session.insert("rose.mary.trace.database.mapper.m01.TraceMapper.insertData", trace);
 				}
@@ -125,10 +126,16 @@ public class TraceService {
 				// trace);
 				// }
 				//
-				// if(loadContents) {
-				// session.insert("rose.mary.trace.database.mapper.m01.TraceMapper.insertContents",
-				// trace);
-				// }
+				
+				// 20230416 
+				// ADD 
+				// 우미건설 요청 (ILIN MI 사이트)
+				// 테스트 미완료
+				if (loadContents && !Util.isEmpty(trace.getData())) {					
+					session.delete("rose.mary.trace.database.mapper.m01.TraceMapper.deleteData", trace);
+					session.insert("rose.mary.trace.database.mapper.m01.TraceMapper.insertData", trace);
+				}
+				
 			}
 
 			bs = session.flushStatements();
